refactor: extract create new url to model url file
